test(editor): split corsa diagnostic filters
